pyproject.toml
[project]
name = "djangocourse"
version = "0.1.0"
description = ""
authors = [
    {name = "Andrey",email = "andreymoises.garcia@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.5,<6.0.0)",
    "django-allauth[socialaccount] (>=65.4.1,<66.0.0)",
    "psycopg (>=3.2.4,<4.0.0)",
    "dj-database-url (>=2.3.0,<3.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^5.0.1"
django-browser-reload = "^1.18.0"

